+++ temba/request_logs/models.py
@@ -106,16 +106,16 @@ class HTTPLog(models.Model):
             HTTPLog.objects.filter(id__in=chunk).delete()
 
     @classmethod
-    def create_from_response(cls, log_type, url, response, classifier=None, channel=None, request_time=None):
-        if classifier is not None:
-            org = classifier.org
-
-        if channel is not None:
-            org = channel.org
+    def create_from_response(
+        cls, log_type, url, response, classifier=None, channel=None, ticketer=None, request_time=None
+    ):
+        org = (classifier or channel or ticketer).org
 
         is_error = response.status_code != 200
         data = dump.dump_response(
-            response, request_prefix=cls.REQUEST_DELIM, response_prefix=cls.RESPONSE_DELIM
+            response,
+            request_prefix=cls.REQUEST_DELIM.encode("utf-8"),
+            response_prefix=cls.RESPONSE_DELIM.encode("utf-8"),
         ).decode("utf-8")
 
         # first build our array of request lines, our last item will also contain our response lines
@@ -134,6 +134,7 @@ class HTTPLog(models.Model):
         return HTTPLog.objects.create(
             classifier=classifier,
             channel=channel,
+            ticketer=ticketer,
             log_type=log_type,
             url=url,
             request=request,
@@ -145,12 +146,8 @@ class HTTPLog(models.Model):
         )
 
     @classmethod
-    def create_from_exception(cls, log_type, url, exception, start, classifier=None, channel=None):
-        if classifier is not None:
-            org = classifier.org
-
-        if channel is not None:
-            org = channel.org
+    def create_from_exception(cls, log_type, url, exception, start, classifier=None, channel=None, ticketer=None):
+        org = (classifier or channel or ticketer).org
 
         data = bytearray()
         prefixes = dump.PrefixSettings(cls.REQUEST_DELIM, cls.RESPONSE_DELIM)
@@ -163,7 +160,8 @@ class HTTPLog(models.Model):
         return HTTPLog.objects.create(
             channel=channel,
             classifier=classifier,
-            log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED,
+            ticketer=ticketer,
+            log_type=log_type,
             url=url,
             request=request,
             response="",